Update on interactable
ALL prefabs and props are in game, correct rig, BC, and scripts attached.

Updated the correct items to use in the exorcism.

Already spawned in items are now removed, and rely on the random spawn location script.

The jug no longer is attached to the "testJug"

"DemonPills" are now an item of a doll and spawns in random locations.

Jug has been moved slightly.

Spawn locations of items have been updated to be in a more ideal place for the player.

Might need help with addressing a door that can be locked and opened with the key, as I do not want to ruin any of the setdress.