Spawned objects is better locations.
Most of the item spawn locations have been moved into more ideal places for the player where it will be a bit more obvious and easier to get to.

Some of the box colliders have been changed.

Props need to be fixed where they only stand-up, unless interfered with in-game.