Update on box colliders and locations of items.
Updated some collision boxes of frames and shelfs to allow props to spawn correctly on these areas.

Changed the locations of the ring and coin to be more ideal for the player to pick up as well there being the recent change to props having a light on them, set in the tutorial branch.